% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_current_history}
\alias{gx_current_history}
\title{gx_current_history}
\usage{
gx_current_history(full = FALSE)
}
\arguments{
\item{full,}{if True, return a list with some history details}
}
\description{
Show the name of the current history
}

