% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_get}
\alias{gx_get}
\title{gx_get}
\usage{
gx_get(file_id, create = FALSE, force = FALSE)
}
\arguments{
\item{file_id,}{ID number}

\item{create,}{if TRUE, create import directory if it does not exist}

\item{force,}{if TRUE, will download the file even if it already exists locally}
}
\description{
Download a dataset from the current Galaxy history by ID #
}

