% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_get_import_directory}
\alias{gx_get_import_directory}
\title{gx_get_import_directory}
\usage{
gx_get_import_directory(create = FALSE)
}
\arguments{
\item{create,}{if TRUE, create import directory if it does not exist}
}
\description{
This function returns the import directory to work with
}

