% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_init}
\alias{gx_init}
\title{gx_init}
\usage{
gx_init(API_KEY = NULL, GALAXY_URL = NULL, HISTORY_ID = NULL,
  IMPORT_DIRECTORY = NULL, TMP_DIRECTORY = NULL)
}
\arguments{
\item{API_KEY,}{to access your Galaxy account}

\item{GALAXY_URL,}{the Galaxy instance to work with}

\item{HISTORY_ID,}{the Galaxy history id to work on}

\item{IMPORT_DIRECTORY,}{default '/tmp/<username>/galaxy_import'}

\item{TMP_DIRECTORY,}{default '/tmp/<username>/galaxy_import/tmp'}
}
\description{
Function that graps any environment/default settings to set for this session
to use.
}

