% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gx_put}
\alias{gx_put}
\title{gx_put}
\usage{
gx_put(filename, file_type = "auto")
}
\arguments{
\item{file_type,}{auto-detect otherwise}

\item{Path}{to file}
}
\description{
This function uploads a dataset to the current Galaxy history
}

