% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_save}
\alias{gx_save}
\title{gx_save}
\usage{
gx_save(session_name = "workspace")
}
\arguments{
\item{session_name,}{default "workspace"}
}
\description{
Save the notebook .RData and .RHistory to Galaxy. Convenience function which wraps save.image and gx_put
}

