% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_get}
\alias{gx_get}
\title{gx_get}
\usage{
gx_get(file_id)
}
\arguments{
\item{file_id,}{Dataset ID number for current history}
}
\description{
Download a dataset from the current Galaxy history by ID #
}

