% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_put}
\alias{gx_put}
\title{gx_put}
\usage{
gx_put(filename, file_type = "auto")
}
\arguments{
\item{filename,}{path to file}

\item{file_type,}{auto-detect otherwise specify string filetype}
}
\description{
This function uploads a dataset to the current Galaxy history
}

